Snowflake QA Automation Scripts/requirements.txt
pandas
numpy
openpyxl
sqlalchemy
pyodbc
snowflake-sqlalchemy
snowflake-connector-python[pandas]